+++ src/System/Console.ts
@@ -1,5 +1,0 @@
-export class Console {
-    public static WriteLine(value: unknown): void {
-        console.log(value ? value.toString() : "");
-    }
-}