+++ src/System/Console/index.ts
@@ -1,0 +1,8 @@
+import { Console } from "./Console";
+import { AnsiConsole } from "./AnsiConsole";
+import { Color } from "./Color";
+import { Rule } from "./Rule";
+import { Panel } from "./Panel";
+import { Table } from "./Table";
+
+export { Console, AnsiConsole, Color, Rule, Panel, Table };